Working AJAX-Flowline Calls
This is a working version of the AJAX flowline calls from the front end.
Three files are included.

views.py and urls.py contain the serverside handling of the url routing, the ajax call, and generating the flowline.

The map_index.html file is a testing ground for the ajax calls only. It includes a working example of an ajax call. It does not draw the lines on the stand in map.